Rework with single bootstrap image, more archs.
This should significantly speedup builds, because the base image build
now runs in a non-emulated environment (except for the package hooks.)

In addition, we now verify the entire build with GPG, no longer skipping
the ARM keyring package. All platforms are built in the same manner.

Because this makes things so simple, we now have builds setup for all
ARM architectures as well!